Products repository
- Added products repository
- Added products management (details, edit, delete)

+++ Models/ViewModels/ProductEditVm.cs
@@ -5,8 +5,8 @@
         public int Id { get; set; }
         public string Name { get; set; }
         public string? Description { get; set; }
-        public int Count { get; set; }
-        public int Price { get; set; }
+        public float Count { get; set; }
+        public float Price { get; set; }
         public string? CategoryName { get; set; }
         public string? AisleName { get; set; }
         public int Row { get; set; }
